Move createDOProviderContainerRuntimeFactory into its own file
